Split discovery and grabbing.
Main.py is the storage server. Worker.py is the grabber server. The
storage server discovers new URLs from the services. The list of these
new URLs is split over the targets in `rsync_targets`. The lists are
synced to the worker. WARCs are send back to the storage server, which
uploads those WARCs to the Internet Archive.
Fix issue #21.
Bump version to 20160308.01.